+++ sources/RI.Utilities/Threading/HeavyThread.cs
@@ -32,7 +32,6 @@ namespace RI.Utilities.Threading
     /// TODO: OnStopped
     /// TODO: HeavyThreadAwaiter (incl. option for thread configuration callback)
     /// TODO: Document exception procedure with list
-    /// TODO: Code example
     public abstract class HeavyThread : IDisposable, ISynchronizable
     {
         #region Constants